+++ qt/scientific_interfaces/ISISReflectometry/test/Runs/RunsPresenterTest.h
@@ -17,7 +17,7 @@
 #include "MantidKernel/ConfigService.h"
 #include "MantidQtWidgets/Common/AlgorithmRuntimeProps.h"
 #include "MantidQtWidgets/Common/Batch/MockJobTreeView.h"
-#include "MantidQtWidgets/Common/MockAlgorithmRunner.h"
+#include "MantidQtWidgets/Common/JobRunners/MockAlgorithmRunner.h"
 #include "MantidQtWidgets/Common/MockProgressableView.h"
 #include "MockRunsView.h"
 #include <cxxtest/TestSuite.h>
